// sdks/typescript/src/generated/api/types/TokenResponse.ts
// This file was auto-generated by Fern from our API Definition.

export interface TokenResponse {
    /** Access token to be used when calling the API. Valid for 12 hours. */
    access_token?: string | undefined;
    /** Type of the token. */
    token_type?: string | undefined;
    /**
     * Refresh token used to obtain new user access tokens without requiring the user to re-authorise. Valid for up to 20 years.
     *
     * Only returned for user access token grant types (`registration_code`, `authorization_code`, `refresh_token`).
     */
    refresh_token?: string | undefined;
    /** Access token expiry time in seconds. */
    expires_in?: number | undefined;
    /** Access token expiration timestamp (UTC). */
    expires_at?: string | undefined;
    /**
     * Refresh token expiry time in seconds.
     *
     * Only returned for user access token grant types (`registration_code`, `authorization_code`, `refresh_token`).
     */
    refresh_token_expires_in?: number | undefined;
    /**
     * Refresh token expiration timestamp (UTC).
     *
     * Only returned for user access token grant types (`registration_code`, `authorization_code`, `refresh_token`).
     */
    refresh_token_expires_at?: string | undefined;
    /** Scope of the token. */
    scope?: string | undefined;
    /**
     * Token creation time in ISO 8601 format.
     *
     * Only returned for user access token grant types (`registration_code`, `authorization_code`, `refresh_token`).
     */
    created_at?: string | undefined;
}